requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
passlib[bcrypt]
PyJWT
httpx
python-jose[cryptography]
google-auth
google-auth-oauthlib
firebase-admin
sentence-transformers
pinecone
pypdf
python-docx
beautifulsoup4
langchain
langchain-community
pymongo
soundfile
pydub
openai
python-multipart
slowapi
openai-whisper
langchain-text-splitters
pandas
unstructured
PyPDF2
openpyxl
redis
redis-om
hiredis
langchain-redis
redis